requirements.txt
python-dotenv>=1.0.1
pytz>=2024.1
requests>=2.32.3
flask>=3.0.0
